metrics/circonus: temporarily remove
The public API was recently updated with many breaking changes, and the
new code is unfortunately completely undocumented. Figuring out how to
work with the new API and reimplementing the tests is a task that will
take some time, and it looks like much of the previous integration work
wouldn't survive the transition. Removing Circonus support until I can
dedicate some time to this.